requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==3.6.4.0
celery==5.2.3
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==3.2.0
django-admin-relation-links==0.2.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-redis==5.4.0
django-restframework==0.0.1
django-reversion==4.0.1
django-timezone-field==4.2.3
djangorestframework==3.15.1
jsonfield==3.1.0
kombu==5.3.7
prompt_toolkit==3.0.47
psycopg2==2.9.3
psycopg2-binary==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
redis==5.0.7
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
gspread
async-timeout
pandas
google-api-core==2.5.0
google-api-python-client==2.123.0
google-auth==2.6.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.5.3
google-cloud-core==2.2.2
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
django-cors-headers
eventlet
django-jsonform
slackclient == 2.9.4 
